Update RTL8188EU driver to v4.3.0.5_12017.20140807: 1. Fix AP mode TXDMA status error 2. Fix 11n logo test item 5.2.16 fail 3. Fix IOT issue with some AP's RA mechanism 4. Fix bug when 2 continual p2p_listen command is given 5. Fix kernel call trace caused by un-initialized spinlock 6. Fix STA packet loss when P2P scan 7. Fix compile warning

+++ rtl8xxx_EU/include/hal_intf.h
@@ -623,8 +623,8 @@ c2h_id_filter rtw_hal_c2h_id_filter_ccx(_adapter *adapter);
 
 s32 rtw_hal_is_disable_sw_channel_plan(PADAPTER padapter);
 
-s32 rtw_hal_macid_sleep(PADAPTER padapter, u32 macid);
-s32 rtw_hal_macid_wakeup(PADAPTER padapter, u32 macid);
+s32 rtw_hal_macid_sleep(PADAPTER padapter, u8 macid);
+s32 rtw_hal_macid_wakeup(PADAPTER padapter, u8 macid);
 
 #ifdef CONFIG_BT_COEXIST
 s32 rtw_hal_fill_h2c_cmd(PADAPTER, u8 ElementID, u32 CmdLen, u8 *pCmdBuffer);
